Add Gemfile and update README accordingly
Fixes #805

The problem is that the PopcornTimeTV custom spec repo does not work with newer version of CocoaPods. Therefore we have to use an old version.

In my solution for this problem I added Gemfile where I specified the cocoapods version to use. this way when we use the `bundle exec pod install` command bundler will use the cocoapods version which actually works.

I added the following changes in this commit:
* added Gemfile
* added cocoapods v1.5.3 to the gemfile
* modified README file accordingly